Y2016Day13 solved.
Searched by taxicab moves first, and then if that fails, allow one non taxicab move and retry.

This approach was very efficient but did not lend itself to part two which required figuring out every reachable tile in 50 steps, so I just rewrote with a new node that counts steps for that part.