Y2016Day21 solved.
To scramble, its a just do it exercise.
To unscramble, just iteratively deduce the inverse command as its fairly simple. e.g. for swap/reverse just do the same command. For rotations turn the rotation around, for move just swap the move indices, and then for rotate based that took the most logic.